challenge10/requirements.txt
google-adk
google-cloud-dlp
GOOGLE_CLOUD_PROJECT=leoy-toolbox

